[DBTools] Relocate from the advanced logbook to separate page
