+++ application/views/dbtools/checkresult.php
